requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Flask-Login==0.6.3
Werkzeug==2.3.8
python-dotenv==1.0.0
authlib==1.2.1
requests==2.31.0
groq==0.9.0
httpx==0.27.0
httpcore==1.0.5
PyPDF2==3.0.1
Pillow==9.5.0
pytesseract==0.3.10
pymongo==4.6.1
razorpay==1.4.1
bcrypt==4.1.2
tiktoken==0.5.2
pymongo[srv]
certifi
dnspython
gunicorn
